+++ core/add_list.php
@@ -7,7 +7,7 @@
  * @author     Dennis Suitters <dennis@diemen.design>
  * @copyright  2014-2019 Diemen Design
  * @license    http://opensource.org/licenses/MIT  MIT License
- * @version    0.2.24
+ * @version    0.2.26-5
  * @link       https://github.com/DiemenDesign/AuroraCMS
  * @notes      This PHP Script is designed to be executed using PHP 7+
  */
@@ -212,6 +212,7 @@ else{
 					'</div>'.
 				'</div>'.
 			'</div>`);'.
+      'window.top.window.toastr["success"]("'.$lh.' added!")'.
 		'</script>';
 	}else echo'<script>window.top.window.toastr["error"]("There was an issue adding the Data!");</script>';
 }